+++ R/medRCT.R
@@ -44,6 +44,10 @@ utils::globalVariables(".SD")
 #'   Only one effect measure can be specified at a time.
 #' @param mcsim An \code{integer} specifying the number of Monte Carlo simulations to perform. The default is 200.
 #' It is recommended to run analysis with no fewer than 200 Monte Carlo simulations.
+#' @param separation_method Method to handle separation, only relevant for binomial (binary outcome) models.
+#'   Options are \code{"brglm"} (Logistic regression models are fitted using bias reduction methods for generalised linear models implemented in the \code{brglm2} package)
+#'   or \code{"discard"} (if separation is detected, the function returns \code{NA}. If this occurs during the main estimation,
+#'   the program stops; if it occurs during bootstrapping, the affected bootstrap samples are discarded).
 #' @param bootstrap A \code{logical} value indicating whether bootstrapping should be performed. If \code{TRUE}
 #'  (default), bootstrapping is conducted using the \code{boot} function from the \code{boot} package.
 #' @param boot_args A \code{list} of arguments for bootstrapping. The default settings are:
@@ -107,6 +111,7 @@ medRCT <- function(
   intervention_type = c("all", "shift_all", "shift_k", "shift_k_order"),
   effect_measure = NULL,
   mcsim = 200,
+  separation_method = "discard",
   bootstrap = TRUE,
   boot_args = list(R = 100, stype = "i", ci.type = "norm"),
   ...
@@ -115,6 +120,13 @@ medRCT <- function(
   choices <- c("all", "shift_all", "shift_k", "shift_k_order")
   idx <- match(intervention_type, choices)
   intervention_type <- choices[idx]
+
+  if (
+    !identical(separation_method, "discard") &&
+      !identical(separation_method, "brglm")
+  ) {
+    stop('separation_method must be either "discard" or "brglm"')
+  }
 
   # set intervention type to shift_k when K==1
   if (
@@ -156,6 +168,12 @@ medRCT <- function(
   if (mcsim < 200) {
     message(
       "Note: It is recommended to run analysis with no fewer than 200 Monte Carlo simulations."
+    )
+  }
+
+  if (separation_method == "brglm") {
+    message(
+      "Note: Logistic regression models are fitted using bias reduction methods for generalised linear models."
     )
   }
 
@@ -213,12 +231,35 @@ medRCT <- function(
   } else if (interactions_XC == "none") {
     interactions_XC <- paste(confounders, collapse = "+")
   } else {
-    interactions_XC <- gsub(exposure, "X", interactions_XC)
+    interactions_XC <- paste0(
+      paste(confounders, collapse = "+"),
+      "+",
+      gsub(exposure, "X", interactions_XC)
+    )
   }
 
   # set R to 1 if bootstrap is not required
   if (bootstrap == FALSE) {
     boot_args$R = 1
+  }
+
+  # test for main estimation
+  main_est_test = medRCT.fun.safe(
+    dat = dat,
+    ind = 1:nrow(dat),
+    first = first,
+    K = K,
+    fam_type = fam_type,
+    mediators = mediators,
+    interactions_XC = interactions_XC,
+    intervention_type = intervention_type,
+    separation_method = separation_method,
+    effect_measure = effect_measure,
+    mcsim = mcsim
+  )
+
+  if (any(is.na(main_est_test))) {
+    stop("Main estimation failed or warned, aborting bootstrap.")
   }
 
   # bootstrap
@@ -234,6 +275,7 @@ medRCT <- function(
     interactions_XC = interactions_XC,
     intervention_type = intervention_type,
     stype = boot_args$stype,
+    separation_method = separation_method,
     R = boot_args$R,
     ...
   )